Use smaller datatype for frame variable

+++ opponent.h
@@ -6,7 +6,7 @@ Sprites sprites;
 
 void drawopponent()
 {
-	int frame = (opp.stance == OppStance::oppStanding) ? 0 : ((opp.stance * 2) - 1 + opp.animationFrame);
+	uint8_t frame = (opp.stance == OppStance::oppStanding) ? 0 : ((opp.stance * 2) - 1 + opp.animationFrame);
 	sprites.drawOverwrite(opp.x, opp.y, opponentImages, frame);
 }
